Feature/Constants refactoring (#1)
* Moved constants in dedicated libraries
* Improved structure of code
