+++ CLion/Air Analyzer/lib/ApiManagement/src/ApiManagementConsts.h
@@ -1,0 +1,4 @@
+const String API_MANAGEMENT_URI_USER_LOGIN =                                "api/user/login";
+const String API_MANAGEMENT_URI_ROOM_CHANGE_STATUS_ACTIVATION =             "api/room/changeStatusActivation";
+const String API_MANAGEMENT_URI_ROOM_API_CHANGE_LOCAL_IP =                  "api/room/changeLocalIP";
+const String API_MANAGEMENT_URI_MEASURE_SET =                               "api/measure/set";
