--- CLion/Air Analyzer/lib/ApiManagement/src/ApiRoute.h
@@ -1,4 +1,0 @@
-const String API_LOGIN =                                        "api/user/login";
-const String API_CHANGE_STATUS_ACTIVATION_ROOM =                "api/room/changeStatusActivation";
-const String API_CHANGE_LOCAL_IP_ROOM =                         "api/room/changeLocalIP";
-const String API_SET_MEASURES =                                 "api/measure/set";
