+++ CLion/Air Analyzer/lib/DatetimeInterval/src/DateTimeIntervalConsts.h
@@ -1,0 +1,2 @@
+const uint8_t DATE_INTERVAL_TIMEOUT_RTC_CHECK_DAY =                     14;
+const uint16_t DATE_INTERVAL_TIMEOUT_NTP_CHECK_MILLISECONDS =           5000; // Timeout to check the datetime from NTP.
