+++ CLion/Air Analyzer/lib/DatetimeInterval/src/DatetimeInterval.h
@@ -1,133 +1,142 @@
- /**
-  * @brief This library allows to manage an interval of time between the actual (stored on RTC module and get with NTP object)
-  * and the next set. 
-  * Copyright (c) 2022 Davide Palladino.
-  * All right reserved.
-  * 
-  * @author Davide Palladino
-  * @contact davidepalladino@hotmail.com
-  * @website https://davidepalladino.github.io/
-  * @version 3.0.3
-  * @date 30th September, 2022
-  * 
-  */
+/**
+ * @file DatetimeInterval.h
+ * @brief Provides functionality to manage time intervals using an RTC module and NTP synchronization.
+ *
+ * This library manages the time intervals by comparing the actual time stored in the RTC module
+ * with the NTP-synchronized time, ensuring periodic updates.
+ *
+ * Copyright (c) 2025 Davide Palladino.
+ * All rights reserved.
+ *
+ * @author Davide Palladino
+ * @contact davidepalladino@hotmail.com
+ * @website https://davidepalladino.github.io/
+ * @version 4.0.0
+ * @date 25th January 2025
+ */
 
 #ifndef DATETIMEINTERVAL_H
     #define DATETIMEINTERVAL_H
 
     #include <Arduino.h>
-
     #include <ESP8266WiFi.h>
-
     #include <NTPClient.h>
     #include <RTClib.h>
+    #include <Wire.h>
 
-    #include <ctime>
+    #include <DateTimeIntervalConsts.h>
 
-    #define TIMEOUT_CHECK_RTC_DAY 14
-    #define TIMEOUT_CHECK_NTP_MILLISECONDS 5000                                            // Timeout to check the datetime from NTP.
-
+    /**
+    * @class DatetimeInterval
+    * @brief Manages time intervals using an RTC module and NTP synchronization.
+    *
+    * This class provides functionality to synchronize time using an NTP server and an RTC module,
+    * allowing periodic updates based on configurable time intervals.
+    */
     class DatetimeInterval {
         public:
-            /** 
-             * @brief This constructor creates this object setting the NTPClient object.
+            /**
+             * @brief Constructs a DatetimeInterval object with NTP client settings.
+             * @param ntpClient Reference to the NTPClient object for time synchronization.
              */
             explicit DatetimeInterval(NTPClient ntpClient);
 
-            /** 
-             * @brief This method initializes the RTC object and sets the next datetime.
-             * @param totalMinuteUpdate Total minutes for every update.
-             * @warning The max minutes will be 240. Every value greater than this will be considered as such.
+            /**
+             * @brief Initializes the RTC object and sets the next update time.
+             * @param totalMinuteUpdate Total minutes between updates (max: 240 minutes).
+             * @warning Values above 240 minutes will be capped.
              */
             void begin(uint8_t totalMinuteUpdate);
 
-            /** 
-             * @brief This method indicates if the actual datetime is greater than next.
-             * @return Boolean value "true" to indicate if there is a variation; else value "false".
+            /**
+             * @brief Checks if the actual datetime exceeds the next scheduled update.
+             * @return True if the update interval has elapsed, false otherwise.
              */
             bool checkDatetime();
 
-            /** 
-             * @brief This method configs the next datetime, based on "timespanDatetime".
+            /**
+             * @brief Configures the next scheduled datetime based on the update interval.
              */
             void configNextDatetime();
 
-            /** 
-             * @brief This method gets the year from the RCT object.
-             * @return Actual year.
+            /**
+             * @brief Retrieves the current year from the RTC module.
+             * @return The current year as a 4-digit integer.
              */
             uint16_t getActualYear();
 
-            /** 
-             * @brief This method gets the month from the RCT object.
-             * @return Actual month.
+            /**
+             * @brief Retrieves the current month from the RTC module.
+             * @return The current month (1-12).
              */
             uint8_t getActualMonth();
 
-            /** 
-             * @brief This method gets the day from the RCT object.
-             * @return Actual day.
+            /**
+             * @brief Retrieves the current day from the RTC module.
+             * @return The current day (1-31).
              */
             uint8_t getActualDay();
 
-            /** 
-             * @brief This method gets the day of week from the RCT object.
-             * @return Actual day of week. Value "0" is about to sunday; "1" for monday.
+            /**
+             * @brief Retrieves the current day of the week from the RTC module.
+             * @return Day of the week (0 = Sunday, 1 = Monday, ..., 6 = Saturday).
              */
             uint8_t getActualDayWeek();
 
-            /** 
-             * @brief TThis method gets the hour from the RCT object.
-             * @return Actual hour.
+            /**
+             * @brief Retrieves the current hour from the RTC module.
+             * @return The current hour (0-23).
              */
             uint8_t getActualHour();
 
-            /** 
-             * @brief This method gets the minute from the RCT object.
-             * @return Actual minute.
+            /**
+             * @brief Retrieves the current minute from the RTC module.
+             * @return The current minute (0-59).
              */
             uint8_t getActualMinute();
 
-            /** 
-             * @brief This method gets the second from the RCT object.
-             * @return Actual second .
+            /**
+             * @brief Retrieves the current second from the RTC module.
+             * @return The current second (0-59).
              */
             uint8_t getActualSecond();
 
             /**
-             * @brief This method gets a struct "tm" from a Datetime object.
-             * @return Struct "tm" with the datetime.
+             * @brief Converts a DateTime object to a tm struct.
+             * @param datetime Reference to the DateTime object.
+             * @return Struct tm representing the datetime.
              */
             struct tm getTmDatetime(DateTime &datetime);
 
             /**
-             * @brief This method gets the timestamp of actual datetime.
-             * @return String with timestamp.
+             * @brief Retrieves the current timestamp as a formatted string.
+             * @return A string representing the current timestamp in format "YYYY-MM-DD HH:MM:SS".
              */
             String getActualTimestamp();
 
         private:
-            NTPClient ntpClient;
-            TimeSpan timespanDatetime;
-            DateTime nextDatetime;
-            RTC_DS3231 rtc;
-            TimeSpan timespanDatetimeRTC;
-            DateTime nextDatetimeRTC;
+            NTPClient ntpClient;             /**< NTP client for time synchronization. */
+            TimeSpan timespanDatetime;       /**< Timespan for next scheduled update. */
+            DateTime nextDatetime;           /**< Stores the next scheduled update time. */
+            RTC_DS3231 rtc;                  /**< RTC module instance. */
+            TimeSpan timespanDatetimeRTC;    /**< RTC-based timespan for updates. */
+            DateTime nextDatetimeRTC;        /**< Next RTC-based update time. */
 
             /**
-             * @brief This method indicates, for RTC object, if the actual datetime is greater than next.
-             * @return Boolean value "true" to indicate if there is a variation; else value "false".
+             * @brief Checks if the current RTC datetime exceeds the next scheduled update.
+             * @return True if the update interval has elapsed, false otherwise.
              */
             bool checkDatetimeRTC();
 
             /**
-             * @brief This method configs the next datetime for RTC object, based on "timespanDatetimeRTC".
+             * @brief Configures the next RTC-based scheduled datetime.
              */
             void configNextDatetimeRTC();
 
             /**
-             * @brief This method updates the RTC module.
+             * @brief Updates the RTC module with the latest synchronized time.
              */
             void updateDatetimeRTC();
     };
-#endif
+
+#endif // DATETIMEINTERVAL_H
