+++ CLion/Air Analyzer/lib/FirmwareUpdateOTA/src/FirmwareUpdateOTA.cpp
@@ -2,19 +2,17 @@
 
 FirmwareUpdateOTA::FirmwareUpdateOTA() { }
 
-void FirmwareUpdateOTA::begin(const String &address, uint16_t port, const String &uri) {
+void FirmwareUpdateOTA::begin(const String &address, uint16_t port) {
     Serial.println("\033[1;92m-------------------- [FIRMWARE] -------------------\033[0m");
 
     this->serverAddress = address;
     this->serverPort = port;
 
-    this->serverUri = uri;
-
     ESPhttpUpdate.rebootOnUpdate(false);
 }
 
 bool FirmwareUpdateOTA::check(const String &version) {
-    t_httpUpdate_return ret = ESPhttpUpdate.update(wifiClient, serverAddress + ":" + String(serverPort) + "/" + serverUri, version);
+    t_httpUpdate_return ret = ESPhttpUpdate.update(wifiClient, serverAddress + ":" + String(serverPort) + "/" + FIRMWARE_UPDATE_OTA_URI_GET_LATEST, version);
     switch (ret) {
         case HTTP_UPDATE_OK:
             Serial.println("\033[1;92m[FIRMWARE UPDATED]\033[0m");
