+++ CLion/Air Analyzer/lib/FirmwareUpdateOTA/src/FirmwareUpdateOTAConsts.h
@@ -1,0 +1,1 @@
+const String FIRMWARE_UPDATE_OTA_URI_GET_LATEST = "api/firmware/getLatest";
