+++ CLion/Air Analyzer/lib/Screen/src/Screen.h
@@ -1,219 +1,229 @@
- /**
-  * @brief This library allows to manage the screen for the Air Analyzer purposes.
-  * The class is an observer where updates own status when is notified by the subject class.
-  * Copyright (c) 2022 Davide Palladino.
-  * All right reserved.
-  * 
-  * @author Davide Palladino
-  * @contact davidepalladino@hotmail.com
-  * @website https://davidepalladino.github.io/
-  * @version 1.5.0
-  * @date 7th October, 2022
-  * 
-  */
+/**
+ * @file Screen.h
+ * @brief Provides functionality to manage the screen for Air Analyzer purposes.
+ *
+ * This library enables screen management as an observer, updating its state when notified by the subject class.
+ *
+ * Copyright (c) 2025 Davide Palladino.
+ * All rights reserved.
+ *
+ * @author Davide Palladino
+ * @contact davidepalladino@hotmail.com
+ * @website https://davidepalladino.github.io/
+ * @version 2.0.0
+ * @date 25th January 2025
+ */
 
 #ifndef SCREEN_H
     #define SCREEN_H
 
     #include <Arduino.h>
-
     #include <U8g2lib.h>
-
     #include <AbstractObserver.h>
     #include <Sensor.h>
 
-    #include "configScreen.h"
-    #include "iconsScreen.h"
+    #include "ScreenConsts.h"
 
     class Sensor;
 
+    /**
+     * @class Screen
+     * @brief Manages screen operations and updates for the Air Analyzer system.
+     *
+     * This class observes the Sensor class and updates the screen accordingly.
+     */
     class Screen : private AbstractObserver {
         friend class AbstractSubject;
 
         public:
-            /** 
-             * @brief This constructor creates the object setting only the subject class and the pins where is
-             *  connected the screen.
-             * @param sensor Subject object that will notify.
-             * @param pinSCL Pin SCL where is connected the screen.
-             * @param pinSDA Pin SDA where i connected the screen.
+            /**
+             * @brief Constructs a Screen object.
+             *
+             * @param sensor Reference to the Sensor object to observe.
+             * @param pinSCL SCL pin connected to the screen (for I2C communication).
+             * @param pinSDA SDA pin connected to the screen (for I2C communication).
+             * // FIXME: Is wrong. `sensor` notify any change.
              */
             Screen(Sensor &sensor, uint8_t pinSCL, uint8_t pinSDA);
 
-            /** 
-             * @brief This method initializes the screen object.
+            /**
+             * @brief Initializes the screen display.
+             *
+             * This method should be called before displaying any content on the screen.
              */
             void begin();
 
             /**
-             * @brief This method sets the room ID.
-             * @param roomNumber Number of room.
+             * @brief Sets the room ID to be displayed on the screen.
+             *
+             * @param roomNumber The ID of the room.
              */
             void setRoomNumber(uint8_t roomNumber);
 
             /**
-             * @brief This method sets the status indicator about the WiFi connection.
-             * @param isConnected Value "true" if there is a WiFi connection; else, value "false".
+             * @brief Sets the Wi-Fi connection status.
+             *
+             * @param isConnected True if connected, false otherwise.
+             * // FIXME: Provide right name.
              */
             void setIsConnected(bool isConnected);
 
             /**
-             * @brief This method gets the status indicator about the WiFi connection.
-             * @return Value "true" if there is a WiFi connection; else, value "false".
+             * @brief Gets the Wi-Fi connection status.
+             *
+             * @return True if connected, false otherwise.
+             * // FIXME: Provide right name.
              */
             bool getIsConnected();
 
             /**
-             * @brief This method sets the status indicator about the update.
-             * @param isUpdated Value "true" if the update has been successful; else, value "false".
+             * @brief Sets the update status of the system.
+             * 
+             * @param isUpdated True if updated, false otherwise.
+             * // FIXME: Provide right name.
              */
             void setIsUpdated(bool isUpdated);
 
             /**
-             * @brief This method gets the status indicator about the update.
-             * @return Value "true" if the update has been successful; else, value "false".
+             * @brief Gets the update status of the system.
+             * 
+             * @return True if updated, false otherwise.
+             * // FIXME: Provide right name.
              */
             bool getIsUpdated();
 
             /**
-             * @brief This method sets the status of screen, if can be rendered or not.
-             * @param isViewable Value "true" if the screen can be rendered.
+             * @brief Sets the screen visibility status.
+             * 
+             * @param isViewable True if the screen should be visible, false otherwise.
+             * // FIXME: Provide right name.
              */
             void setIsViewable(bool isViewable);
 
             /**
-             * @brief This method gets the status of screen, if can be rendered or not.
-             * @return
+             * @brief Gets the screen visibility status.
+             * 
+             * @return True if the screen is viewable, false otherwise.
+             * // FIXME: Provide right name.
              */
             bool getIsViewable();
 
-
-
-        // INSTALLATION and CONFIGURATION VIEWS
+        
+            // VIEWS for installation and configuration
             /**
-             * @brief This methods shows an installation page about the room ID.
-             * @param messages Array of three messages to visualize.
+             * @brief Displays the installation page for room ID setup.
+             * 
+             * @param messages An array of messages to display.
              */
             void showInstallationRoomIDPage(const String messages[3]);
 
             /**
-             * @brief This methods shows an installation page about the WiFi.
-             * @param messages Array of six messages to visualize.
-             * @param result Type of rows to visualize: "0" for three rows considering the first, second and third message;
-             *  "1" considering the fourth and fifth messages ; "2" considering the sixth message.
+             * @brief Displays the installation page for Wi-Fi setup.
+             * 
+             * @param messages An array of messages to display.
+             * @param result A status code representing the Wi-Fi connection result.
              */
             void showInstallationWiFiPage(const String messages[6], uint8_t result);
 
             /**
-             * @brief This methods shows an configuration page about the update two version.
-             * @param messages Array of three messages to visualize. To the last message will be append the "localIP" value.
-             * @param localIP Value selected and that must be show.
+             * @brief Displays the firmware upgrade version screen.
+             * 
+             * @param messages Pointer to an array of messages.
+             * @param localIP The local IP address to display.
              */
             void showUpgradeVersionThreePage(const String *messages, const String &localIP);
 
 
-
-            // ORDINARY VIEWS
+            // VIEWS ordinary
             /**
-             * @brief This method shows the brand and the version of the system.
-             * @param version Version of the system.
+             * @brief Displays the brand logo along with the firmware version.
+             * 
+             * @param version The firmware version string to be displayed.
              */
-             void showBrand(const String &version);
+            void showBrand(const String &version);
 
             /**
-             * @brief This method shows a page with an loading bar and a description message.
-             * @param message The specific message to visualize.
-             * @param percentage Percentage to fill the loading bar.
+             * @brief Displays a loading page with progress.
+             * 
+             * @param message The loading message to display.
+             * @param percentage The loading percentage value.
              */
             void showLoadingPage(const String &message, float percentage);
 
-            /**
-             * @brief This method shows the main page using the sensor object (to visualize the values) the room ID
-             *  and the status of connection and update.
-             */
+            /** @brief Displays the main screen page with sensor data. */
             void showMainPage();
 
-            /**
-             * @brief This method clear the page with all contents.
-             */
+            /** @brief Clears the screen display. */
             void clear();
 
             /**
-             * @brief This method shows a simple page with a simple row of message at the center.
-             * @param message The specific row message to visualize.
-             * @warning The row of message shall be of maximum 18 character. Other characters may be lost at right of screen.
+             * @brief Displays a message on the screen.
+             * 
+             * @param message The message to display.
              */
             void showMessagePage(const String &message);
 
             /**
-             * @brief This method shows a simple page with two simple rows of messages at the center.
-             * @param messages The two messages to visualize.
-             * @warning The single row of message shall be of maximum 18 character. Other characters may be lost at right of screen.
+             * @brief Displays two messages on the screen.
+             * 
+             * @param messages An array containing two messages to display.
              */
             void showMessagePage(const String messages[2]);
-            
-        private:
-            U8G2_SSD1306_128X32_UNIVISION_F_HW_I2C *screen;
-            Sensor &sensor;
-            uint8_t roomNumber;
-            bool isConnected;
-            bool isUpdated;
-            bool isViewable;
 
-            /** 
-             * @brief This method draws the brand.
-             */
+        private:
+            U8G2_SSD1306_128X32_UNIVISION_F_HW_I2C *screen;         /**< Pointer to the screen object. */
+            Sensor &sensor;                                         /**< Reference to the Sensor object. */
+            uint8_t roomNumber;                                     /**< Stores the room number. */
+            bool isConnected;                                       /**< Stores the Wi-Fi connection status. */
+            bool isUpdated;                                         /**< Stores the update status. */
+            bool isViewable;                                        /**< Stores the screen visibility status. */
+
+            /** @brief Draws the brand logo on the screen. */
             void drawBrand();
 
-            /** 
-             * @brief This method draws the version of the system.
-             * @param version Version of the system.
+            /**
+             * @brief Draws the firmware version on the screen.
+             * 
+             * @param version The firmware version string to display.
              */
             void drawVersion(const String &version);
 
-            /** 
-             * @brief This method draws the bar and the internal fill considering the percenage.
-             * @param percentage Percentage to fill the loading bar.
+            /**
+             * @brief Draws a progress bar on the screen.
+             * 
+             * @param percentage The loading percentage to be displayed.
              */
             void drawBar(float percentage);
 
             /**
-             * @brief This method draws the single message considering the "x" and "y" positions.
-             * @param positionX X position.
-             * @param positionY Y position.
-             * @param message The specific row message to visualize.
-             * @warning The row of message shall be of maximum 18 character. Other characters may be lost at right of screen.
+             * @brief Draws a message at the specified position on the screen.
+             * 
+             * @param positionX X-coordinate of the message.
+             * @param positionY Y-coordinate of the message.
+             * @param message The message to display.
              */
             void drawMessage(uint8_t positionX, uint8_t positionY, const String &message);
 
-            /**
-             * @brief This method draws a circle with the room ID inside.
-             */
+            /** @brief Draws the room ID on the screen. */
             void drawRoomID();
 
-            /**
-             * @brief This method draws the icon of temperature and the value read.
-             */
+            /** @brief Draws the current temperature on the screen. */
             void drawTemperature();
 
-            /**
-             * @brief This method draws the icon of humidity and the value read.
-             */
+            /** @brief Draws the current humidity on the screen. */
             void drawHumidity();
 
-            /**
-             * @brief This method draws the status icon if the connection has been successful.
-             */
+            /** @brief Draws the Wi-Fi connection status icon. */
             void drawWiFiStatus();
 
-            /**
-             * @brief This method draws the status icon if the update has not been successful.
-             */
+            /** @brief Draws the update status icon. */
             void drawUpdateStatus();
 
-            /** 
-             * @brief This method update the main screen.
+            /**
+             * @brief Updates the screen content based on the observed changes.
+             * // FIXME: Is wrong. update receives values.
              */
             void update();
     };
-#endif
+
+#endif // SCREEN_H
