+++ CLion/Air Analyzer/lib/Screen/src/ScreenConsts.h
@@ -38,3 +38,13 @@ const String messagePageInstallationCompleteMessages[2] =               {"INSTAL
 const String messagePageSearchingMessage =                              {"WPS CONNECTION..."};
 const String messagePageSuccessfulMessage =                             {"CONNECTED!"};
 const String messagePageErrorMessages[2] =                              {"NO NEW CONNECTION", "REVEALED!"};
+
+static unsigned char logoTemperature[] PROGMEM = {
+    0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01,
+    0xe0, 0x01, 0xe0, 0x01, 0xf0, 0x03, 0xf8, 0x07, 0xf8, 0x07, 0xf8, 0x07,
+    0xf0, 0x03, 0xe0, 0x01 };
+
+static unsigned char logoHumidity[] PROGMEM = {
+    0xc0, 0x00, 0xe0, 0x01, 0xf0, 0x03, 0xf8, 0x07, 0xf8, 0x07, 0xfc, 0x0f,
+    0xfc, 0x0f, 0xfc, 0x0f, 0xfe, 0x1f, 0xfe, 0x1f, 0xfc, 0x0f, 0xfc, 0x0f,
+    0xf8, 0x07, 0xf0, 0x03 };
