--- CLion/Air Analyzer/lib/Screen/src/iconsScreen.h
@@ -1,9 +1,0 @@
-static unsigned char logoTemperature[] PROGMEM = {
-    0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01, 0xe0, 0x01,
-    0xe0, 0x01, 0xe0, 0x01, 0xf0, 0x03, 0xf8, 0x07, 0xf8, 0x07, 0xf8, 0x07,
-    0xf0, 0x03, 0xe0, 0x01 };
-
-static unsigned char logoHumidity[] PROGMEM = {
-    0xc0, 0x00, 0xe0, 0x01, 0xf0, 0x03, 0xf8, 0x07, 0xf8, 0x07, 0xfc, 0x0f,
-    0xfc, 0x0f, 0xfc, 0x0f, 0xfe, 0x1f, 0xfe, 0x1f, 0xfc, 0x0f, 0xfc, 0x0f,
-    0xf8, 0x07, 0xf0, 0x03 };
