+++ CLion/Air Analyzer/lib/Sensor/src/SensorConsts.h
@@ -1,0 +1,1 @@
+const uint16_t TIMEOUT_READ_HDC = 1000;
