+++ CLion/Air Analyzer/lib/ServerSocketJSON/src/ServerSocketJSON.h
@@ -1,22 +1,23 @@
 /**
- * @brief This library allows the device to work like server socket, exchanging JSON messages.
- * The structure of JSON message will have to be:
+ * @file ServerSocketJSON.h
+ * @brief This library allows the device to function as a server socket, exchanging JSON messages.
+ *
+ * The structure of the JSON message is expected to be:
  * {
  *  "Request code": <number between 0 and 255>,
  *  "message": { <JSON nested> }
  * }
  *
- * Copyright (c) 2022 Davide Palladino.
- * All right reserved.
+ * Copyright (c) 2025 Davide Palladino.
+ * All rights reserved.
  *
  * @author Davide Palladino
  * @contact davidepalladino@hotmail.com
  * @website https://davidepalladino.github.io/
- * @version 1.0.1
- * @date 23rd January 2025
+ * @version 2.0.1
+ * @date 25th January 2025
  *
  * @include ArduinoJson v6.18.2.
- *
  */
 
 #ifndef SERVERSOCKETJSON_H
@@ -25,89 +26,107 @@
     #include <Arduino.h>
     #include <ESP8266WiFi.h>
 
+    #include "ServerSocketJSONConsts.h"
+
     #ifndef ARDUINO_JSON
         #include "ArduinoJson.h"
     #endif
 
-    constexpr uint32_t PORT_SERVER_SOCKET = 60000;
-    constexpr uint16_t SIZE_JSON_DOCUMENT = 512;
-    constexpr uint8_t SIZE_CREDENTIAL_USERNAME = 20;
-    constexpr uint8_t SIZE_CREDENTIAL_PASSWORD = 64;
-    const String FIELD_REQUEST_CODE = "request_code";
-    const String FIELD_MESSAGE = "message";
-    const String FIELD_MESSAGE_USERNAME = "username";
-    const String FIELD_MESSAGE_PASSWORD = "password";
-
+    /**
+     * @class ServerSocketJSON
+     * @brief A class that represents a server socket for handling JSON messages with a client.
+     *
+     * This class enables the device to act as a server, receiving and sending JSON messages via Wi-Fi.
+     * It provides methods for managing clients, checking server/client status, and sending/receiving JSON messages.
+     */
     class ServerSocketJSON {
         public:
             /**
-             * @brief This constructor provides to initialize the object.
+             * @brief Constructor to initialize the server socket.
              */
             ServerSocketJSON();
 
             /**
-             * @brief This method opens the server socket on a specific port.
-             * @param port Number of port where the clients will be able to connect.
-             * @return Value "true" if the server is opened; else, value "false". In this case the reason might be a not
-             *  connection of WiFi, or the server already connected.
+             * @brief Initializes the server socket and begins listening for clients.
+             *
+             * This method starts the server socket on the specified port and ensures Wi-Fi is connected.
+             *
+             * @param port The port to listen for incoming connections.
+             * @return True if the server was successfully started, false if Wi-Fi is not connected.
              */
             bool begin(uint16_t port);
 
             /**
-             * @brief This method closes the server and client connections.
+             * @brief Closes the server and any active client connection.
              */
             void end();
 
             /**
-             * @brief This method provides to attach the client, if someone is not already attached.
-             * @return Value "true" if the attaching has been successful; else value "false" to indicate that the attaching
-             *  has been not successful for several reason like the server and/or WiFi not connected, or there is another client attached.
+             * @brief Attaches a client to the server.
+             *
+             * If no client is currently attached, this method establishes a connection with an available client.
+             *
+             * @return True if a client was successfully attached, false otherwise.
              */
             bool attachClient();
 
             /**
-             * @brief This method provides to detach the client connected.
-             * @return Value "true" if the detaching has been successful; else value "false" to indicate that the detaching
-             *  has been not successful because there is not client attached.
+             * @brief Detaches the currently connected client.
+             *
+             * Disconnects the active client, freeing up the server to accept new connections.
+             *
+             * @return True if a client was successfully detached, false if no client is attached.
              */
-
             bool detachClient();
 
             /**
-             * @brief This method provides to check the server connection.
-             * @return Value "true" if the server is connected; else value "false" to indicate that the server is not connected
-             *  because there is not connection of WiFi.
+             * @brief Checks if the server is actively running.
+             *
+             * Determines whether the server is active and ready to accept client connections.
+             *
+             * @return True if the server is running, false otherwise.
              */
             bool isConnected();
 
             /**
-             * @brief This method provides to check if there is a client connected.
-             * @return Value "true" if there is a client connected; else, value "false".
+             * @brief Checks if a client is currently connected.
+             *
+             * @return True if a client is attached, false otherwise.
              */
             bool isAttached();
 
             /**
-             * @brief This method provides to listen a client JSON message.
-             * @return Value of "Request code" contained on JSON message.
+             * @brief Listens for an incoming JSON message from the client.
+             *
+             * This method blocks until a message is received, then deserializes it into a JSON object.
+             *
+             * @return A status code indicating the message reception outcome.
              */
             uint8_t listen();
 
             /**
-             * @brief This method provides to send a specific JSON message to client.
-             * @param message JSON Message to send to client.
-             * @return Value "true" if the JSON message has been sent; else, value "false" to indicate a server and/or WiFi not connected.
+             * @brief Sends a JSON message to the connected client.
+             *
+             * Serializes and transmits a JSON-formatted message to the client.
+             *
+             * @param message The JSON message to be sent in string format.
+             * @return True if the message was successfully sent, false otherwise.
              */
-            bool speak(String &message);
+            bool speak(const String &message);
 
             /**
-             * @brief This method provides to get the last JSON message from client, to deserialize.
-             * @return Field "message" of JSON message to deserialize.
+             * @brief Retrieves the last received JSON message in serialized form.
+             *
+             * This method provides access to the most recently received JSON message.
+             *
+             * @return A string containing the raw JSON message received from the client.
              */
             String getJsonRequestSerialized();
 
         private:
-            WiFiServer *server;
-            WiFiClient client;
-            String jsonRequestSerialized;
+            WiFiServer *server;                 ///< Pointer to the WiFiServer instance for handling client connections.
+            WiFiClient client;                  ///< WiFiClient instance to interact with the connected client.
+            String jsonRequestSerialized;       ///< Holds the last received serialized JSON request.
     };
+
 #endif
