+++ CLion/Air Analyzer/lib/ServerSocketJSON/src/ServerSocketJSONConsts.h
@@ -1,0 +1,6 @@
+constexpr uint8_t SERVER_SOCKET_SIZE_USERNAME =         20;
+constexpr uint8_t SERVER_SOCKET_SIZE_PASSWORD =         64;
+const String SERVER_SOCKET_FIELD_REQUEST_CODE =         "request_code";
+const String SERVER_SOCKET_FIELD_MESSAGE =              "message";
+const String SERVER_SOCKET_FIELD_MESSAGE_USERNAME =     "username";
+const String SERVER_SOCKET_FIELD_MESSAGE_PASSWORD =     "password";
